RT-DETRv4/requirements.txt
absl-py==2.4.0
accelerate==1.13.0
annotated-doc==0.0.4
anyio==4.13.0
calflops==0.3.2
certifi==2026.2.25
click==8.3.2
faster-coco-eval==1.7.2
filelock==3.25.2
fsspec==2026.3.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.9.1
idna==3.11
jinja2==3.1.6
lightning-utilities==0.15.3
markdown-it-py==4.0.0
markdown==3.10.2
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
protobuf==7.34.1
psutil==7.2.2
pygments==2.20.0
pyyaml==6.0.3
regex==2026.4.4
rich==14.3.3
safetensors==0.7.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
sympy==1.13.1
tensorboard-data-server==0.7.2
tensorboard==2.20.0
termcolor==3.3.0
tokenizers==0.22.2
torch==2.5.1+cu124
torchmetrics==1.9.0
torchvision==0.20.1+cu124
tqdm==4.67.3
transformers==5.5.0
triton==3.1.0
typer==0.24.1
typing_extensions==4.15.0
werkzeug==3.1.8
